backend/requirements.txt
fastapi
uvicorn
groq
sqlalchemy
aiofiles
python-dotenv
pytest
